+++ pkg/webui/console/store/actions/gateways.js
@@ -13,6 +13,12 @@
 // limitations under the License.
 
 import createRequestActions from '@ttn-lw/lib/store/actions/create-request-actions'
+import {
+  createPaginationRequestActions,
+  createPaginationBaseActionType,
+  createPaginationDeleteBaseActionType,
+  createPaginationDeleteActions,
+} from '@ttn-lw/lib/store/actions/pagination'
 
 import createGetRightsListRequestActions, { createGetRightsListActionType } from './rights'
 import {
@@ -29,12 +35,6 @@ import {
   createGetEventMessageSuccessActionType,
   getEventMessageSuccess,
 } from './events'
-import {
-  createPaginationRequestActions,
-  createPaginationBaseActionType,
-  createPaginationDeleteBaseActionType,
-  createPaginationDeleteActions,
-} from './pagination'
 
 export const SHARED_NAME = 'GATEWAY'
