+++ pkg/webui/console/store/selectors/collaborators.js
@@ -12,17 +12,17 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+import { createFetchingSelector } from '@ttn-lw/lib/store/selectors/fetching'
+import { createErrorSelector } from '@ttn-lw/lib/store/selectors/error'
+import {
+  createPaginationIdsSelectorByEntity,
+  createPaginationTotalCountSelectorByEntity,
+} from '@ttn-lw/lib/store/selectors/pagination'
+
 import {
   GET_COLLABORATOR_BASE,
   GET_COLLABORATORS_LIST_BASE,
 } from '@console/store/actions/collaborators'
-
-import { createFetchingSelector } from './fetching'
-import { createErrorSelector } from './error'
-import {
-  createPaginationIdsSelectorByEntity,
-  createPaginationTotalCountSelectorByEntity,
-} from './pagination'
 
 const ENTITY = 'collaborators'
