+++ pkg/webui/console/store/selectors/configuration.js
@@ -12,13 +12,13 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+import { createFetchingSelector } from '@ttn-lw/lib/store/selectors/fetching'
+import { createErrorSelector } from '@ttn-lw/lib/store/selectors/error'
+
 import {
   GET_NS_FREQUENCY_PLANS_BASE,
   GET_GS_FREQUENCY_PLANS_BASE,
 } from '@console/store/actions/configuration'
-
-import { createFetchingSelector } from './fetching'
-import { createErrorSelector } from './error'
 
 const selectConfigurationStore = state => state.configuration
