+++ pkg/webui/console/store/selectors/webhook-formats.js
@@ -12,10 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import { GET_WEBHOOK_FORMATS_BASE } from '@console/store/actions/webhook-formats'
+import { createFetchingSelector } from '@ttn-lw/lib/store/selectors/fetching'
+import { createErrorSelector } from '@ttn-lw/lib/store/selectors/error'
 
-import { createFetchingSelector } from './fetching'
-import { createErrorSelector } from './error'
+import { GET_WEBHOOK_FORMATS_BASE } from '@console/store/actions/webhook-formats'
 
 const selectWebhookFormatsStore = state => state.webhookFormats
