+++ pkg/webui/console/store/selectors/webhooks.js
@@ -12,10 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import { GET_WEBHOOK_BASE, GET_WEBHOOKS_LIST_BASE } from '@console/store/actions/webhooks'
+import { createFetchingSelector } from '@ttn-lw/lib/store/selectors/fetching'
+import { createErrorSelector } from '@ttn-lw/lib/store/selectors/error'
 
-import { createFetchingSelector } from './fetching'
-import { createErrorSelector } from './error'
+import { GET_WEBHOOK_BASE, GET_WEBHOOKS_LIST_BASE } from '@console/store/actions/webhooks'
 
 const selectWebhookStore = state => state.webhooks
