+++ pkg/webui/lib/store/logics/create-request-logic.js
@@ -76,7 +76,8 @@ const createRequestLogic = function(
           _resolve(res)
         }
       } catch (e) {
-        error(e) // Log the error if in development mode.
+        // Log the error when in development mode
+        error(e)
 
         if (isUnauthenticatedError(e)) {
           // If there was an unauthenticated error, the access token is not
