+++ pkg/webui/lib/store/reducers/ui/error.js
