+++ pkg/webui/lib/store/selectors/pagination.js
